+++ src/globals.jl
@@ -37,9 +37,7 @@ const Globals = GlobalsType()
 
 function getproperty(::GlobalsType, name::Symbol)
     v = _ValueGlobalVariable(name)
-    if v === C_NULL
-        error("GAP variable $name not bound")
-    end
+    v === C_NULL && error("GAP variable $name not bound")
     return _GAP_TO_JULIA(v)
 end
 
@@ -48,11 +46,27 @@ function hasproperty(::GlobalsType, name::Symbol)
 end
 
 function setproperty!(::GlobalsType, name::Symbol, val::Any)
-    if !CanAssignGlobalVariable(name)
-        error("cannot assing to $name in GAP")
-    end
+    CanAssignGlobalVariable(name) || error("cannot assing to $name in GAP")
     tmp = (val === nothing) ? C_NULL : _JULIA_TO_GAP(val)
     _AssignGlobalVariable(name, tmp)
 end
 
 propertynames(::GlobalsType) = Vector{Symbol}(Globals.NamesGVars())
+
+
+# HACK to get tab completion to work for GAP globals accessed via GAP.Globals;
+# e.g. if the REPL already shows `GAP.Globals.MTX.Is` and the user presses
+# TAB, they should be shown a list of members of the GAP global variable `MTX`
+# (which is a record) starting with `Is`. The easy part for supporting this is
+# to implement `propertynames` for `GapObj` (at least those which are GAP
+# records). Unfortunately that's not quite enough; we also have add methods
+# for the `get_value` method below
+import REPL.REPLCompletions: get_value
+function get_value(sym::Symbol, ::GAP.GlobalsType)
+    v = _ValueGlobalVariable(sym)
+    v === C_NULL && return (nothing, false)
+    return (_GAP_TO_JULIA(v), true)
+end
+get_value(sym::QuoteNode, fn::GAP.GlobalsType) = get_value(sym.value, fn)
+
+propertynames(r::GapObj) = Wrappers.IsRecord(r) ? Vector{Symbol}(Wrappers.RecNames(r)) : Vector{Symbol}()